login Test/requirements.txt

en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz#sha256=06c80936324012d1223291d2af41a5229e746dc2dee8fe31a532666ee3d18aaa
exceptiongroup==1.1.3
executing==2.0.0
fastjsonschema==2.18.1
Flask==3.0.3
gitdb==4.0.11
GitPython==3.1.43
idna==3.4
ipykernel==6.26.0
ipython==8.16.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.5.0
jupyter_core==5.4.0
jupyter_server==2.9.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
markdown-it-py==3.0.0
matplotlib==3.6.2
matplotlib-inline==0.1.6
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.23.3
opencv-python==4.6.0.66
overrides==7.4.0
packaging==22.0
pafy==0.5.5
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pdfminer.six==20231228
pdfminer3==2018.12.3.0
pdfplumber==0.11.2
phonenumbers==8.13.40
pickleshare==0.7.5
Pillow==9.3.0
plac==1.1.3
platformdirs==3.11.0
plotly==5.22.0
preshed==3.0.9
prometheus-client==0.17.1
prompt-toolkit==3.0.39
pyarrow==17.0.0
pycparser==2.21
Pygments==2.16.1
pyparsing==3.0.9
PyPDF2==3.0.1
pypdfium2==4.30.0
pyresparser==1.0.6
pyrsistent==0.20.0
python-dateutil==2.8.2
python-json-logger==2.0.7
regex==2024.5.15
requests==2.31.0
resume-parser==0.8.4
spacy==2.3.9
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==1.0.7
stack-data==0.6.3
stemming==1.0.1
streamlit
streamlit-tags==1.2.8
streamlit_option_menu
tabulate==0.9.0
tornado==6.3.3
typer==0.12.3
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
uri-template==1.3.0
urllib3==2.0.7
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
